refactor: update response model names and module name
